// src/app/pages/live/model/battlegroundevent.ts
export interface BattleGroundEvent {
    eventTime: number;
    eventType: BattleGroundEventType;
}

export enum BetType {
	VALUE = 'value',
	PERCENTAGE = 'percentage',
	HALF = 'half',
	ALLIN = 'allin',
	FLOOR = 'floor',
	RANDOM = 'random'
}

export enum Gender {
	MALE = 'male',
	FEMALE = 'female',
	MONSTER = 'monster'
}

export enum BattleGroundEventType {
	FIGHT_BEGINS = 'FIGHT_BEGINS',
	BETTING_BEGINS = 'BETTING_BEGINS',
	BETTING_ENDS = 'BETTING_ENDS',
	BATTLE_ENDS = 'BATTLE_ENDS',
	BALANCE = 'BALANCE',
	OTHER_PLAYER_BALANCE = 'OTHER_PLAYER_BALANCE',
	LEVEL_UP = 'LEVEL_UP',
	BET = 'BET',
	WIN = 'WIN',
	LOSS = 'LOSS',
	MATCH_INFO = 'MATCH_INFO',
	TEAM_INFO = 'TEAM_INFO',
	EXP = 'EXP ',
	OTHER_PLAYER_EXP = 'OTHER_PLAYER_EXP',
	UNIT_INFO = 'UNIT_INFO',
	BAD_BET = 'BAD_BET ',
	ALLEGIANCE = 'ALLEGIANCE',
	SKILL_DROP = 'SKILL_DROP',
	PLAYER_SKILL = 'PLAYER_SKILL',
	BUY_SKILL = 'BUY_SKILL',
	RESULT = 'RESULT',
	PORTRAIT = 'PORTRAIT',
	SKILL_WIN = 'SKILL_WIN',
	BET_INFO = 'BET_INFO',
	PRESTIGE_SKILLS = 'PRESTIGE_SKILLS',
	LAST_ACTIVE = 'LAST_ACTIVE',
	GIFT_SKILL = 'GIFT_SKILL',
	GLOBAL_GIL_COUNT_UPDATE = 'GLOBAL_GIL_COUNT_UPDATE',
	PRESTIGE_ASCENSION = 'PRESTIGE_ASCENSION',
	MUSIC = 'MUSIC',
	RISER_SKILL_WIN = 'RISER_SKILL_WIN',
	TOURNAMENT_STATUS_UPDATE_EVENT = 'TOURNAMENT_STATUS_UPDATE_EVENT',
	FIGHT_ENTRY = 'FIGHT_ENTRY',
	INVALID_FIGHT_ENTRY_COMBINATION = 'INVALID_FIGHT_ENTRY_COMBINATION',
	INVALID_FIGHT_ENTRY_CLASS = 'INVALID_FIGHT_ENTRY_CLASS',
	UNOWNED_SKILL = 'UNOWNED_SKILL',
	SKILL_ON_COOLDOWN = 'SKILL_ON_COOLDOWN',
	OTHER_PLAYER_UNOWNED_SKILL = 'OTHER_PLAYER_UNOWNED_SKILL',
	OTHER_PLAYER_INVALID_FIGHT_COMBINATION = 'OTHER_PLAYER_INVALID_FIGHT_COMBINATION',
	OTHER_PLAYER_INVALID_FIGHT_CLASS = 'OTHER_PLAYER_INVALID_FIGHT_CLASS',
	OTHER_PLAYER_SKILL_ON_COOLDOWN = 'OTHER_PLAYER_SKILL_ON_COOLDOWN',
	PING = 'PING',
	DONT_FIGHT = 'DONT_FIGHT',
	BUY_SKILL_RANDOM = 'BUY_SKILL_RANDOM'
}